+++ app/lib/ui/components/button/button_demo_screen.dart
@@ -23,10 +23,10 @@ import 'package:ouds_flutter_demo/ui/components/button/button_customization_util
 import 'package:ouds_flutter_demo/ui/components/button/button_enum.dart';
 import 'package:ouds_flutter_demo/ui/theme/theme_controller.dart';
 import 'package:ouds_flutter_demo/ui/utilities/code.dart';
+import 'package:ouds_flutter_demo/ui/utilities/component_screen_header.dart';
 import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_chips.dart';
 import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_section.dart';
 import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_textfield.dart';
-import 'package:ouds_flutter_demo/ui/utilities/detail_screen_header.dart';
 import 'package:provider/provider.dart';
 
 /// This screen displays a button demo and allows customization of button properties
@@ -79,7 +79,7 @@ class _BodyState extends State<_Body> {
   @override
   Widget build(BuildContext context) {
     ThemeController? themeController = Provider.of<ThemeController>(context, listen: false);
-    return DetailScreenDescription(
+    return ComponentScreenHeader(
       description: context.l10n.app_components_button_description_text,
       widget: Column(
         children: [
@@ -115,7 +115,7 @@ class _ButtonDemoState extends State<_ButtonDemo> {
     });
 
     return OudsColoredBox(
-      color: customizationState?.hasOnColoredBox == true ? OudsColoredBoxColor.brandPrimary : null,
+      color: customizationState?.hasOnColoredBox == true ? OudsColoredBoxColor.brandPrimary : OudsColoredBoxColor.statusNeutralMuted,
       child: OudsButton(
         label: ButtonCustomizationUtils.getText(customizationState),
         icon: ButtonCustomizationUtils.getIcon(customizationState),
@@ -142,6 +142,7 @@ class _CustomizationContentState extends State<_CustomizationContent> {
   @override
   Widget build(BuildContext context) {
     final ButtonCustomizationState? customizationState = ButtonCustomization.of(context);
+    final labelFocus = FocusNode();
 
     return CustomizableSection(
       children: [
@@ -205,6 +206,8 @@ class _CustomizationContentState extends State<_CustomizationContent> {
         CustomizableTextField(
           title: context.l10n.app_components_common_text_label,
           text: customizationState.textValue,
+          focusNode: labelFocus,
+          fieldType: FieldType.label,
         )
       ],
     );